Fixed bug where an AreaId between 64 and 255 wouldn't have the correct costs during pathfinding. Split Poly's area and type into two separate variables to allow the full range of AreaIds. Other minor refactors using StyleCop.

+++ SharpNav/NavMeshBuilder.cs
@@ -252,7 +252,7 @@ namespace SharpNav
 				navPolys[i] = new Poly();
 				navPolys[i].vertCount = 0;
 				navPolys[i].flags = parameters.polys[i].Flags;
-				navPolys[i].Area = (int)parameters.polys[i].Area;
+				navPolys[i].Area = parameters.polys[i].Area;
 				navPolys[i].PolyType = PolygonType.Ground;
 				navPolys[i].verts = new int[nvp];
 				navPolys[i].neis = new int[nvp];
